requirements.txt
asgiref==3.2.3
certifi==2020.6.20
dj-database-url==0.5.0
Django==3.0.3
django-ckeditor==5.9.0
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.0.4
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
wincertstore==0.2
